fix: app 전체에서 포그라운드 작동하게 코드 변경
