+++ src/App.js
@@ -1,8 +1,11 @@
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import { ReactQueryDevtools } from "@tanstack/react-query-devtools";
+import { onMessage } from "firebase/messaging";
+import { useEffect } from "react";
 import { Navigate, Route, Routes } from "react-router-dom";
 import Layout from "./components/Layout";
 import ScrollToTop from "./components/ScrollToTop";
+import { messaging } from "./firebase";
 import Admin from "./pages/AdminPages/Admin";
 import BoardEdit from "./pages/AdminPages/BoardEdit";
 import BoardPost from "./pages/AdminPages/BoardPost";
@@ -20,8 +23,53 @@ import PwChange from "./pages/PwChange";
 import Quest from "./pages/Quest";
 import SplashScreen from "./pages/SplashScreen";
 
+if (Notification.permission !== "granted") {
+  await Notification.requestPermission();
+}
+
 function App() {
   const queryClient = new QueryClient();
+
+  // 포그라운드 메시지 수신
+  useEffect(() => {
+    const unsubscribe = onMessage(messaging, (payload) => {
+      console.log("Foreground message received:", payload);
+
+      try {
+        const title = payload.notification?.title || "알림";
+        const body = payload.notification?.body || "";
+        const redirectPath = payload.data?.redirectPath || "/";
+
+        console.log("title:", title);
+        console.log("body:", body);
+        console.log("redirectPath:", redirectPath);
+        console.log("Notification 생성 시도");
+
+        const now = new Date().toLocaleTimeString();
+        setTimeout(() => {
+          const notification = new Notification(`${title} (${now})`, {
+            body,
+            icon: `${window.location.origin}/dohigh.png`,
+            requireInteraction: true,
+            data: { redirectPath },
+          });
+
+          notification.onclick = (event) => {
+            const path = event.target?.data?.redirectPath;
+            console.log("알림 클릭됨 → 이동:", path);
+            if (path) {
+              window.focus();
+              window.location.href = path;
+            }
+          };
+        }, 1000);
+      } catch (err) {
+        console.error("알림 생성 실패:", err);
+      }
+    });
+
+    return () => unsubscribe();
+  }, []);
 
   return (
     <QueryClientProvider client={queryClient}>
